Running on Ubuntu VM on Mac
It took some help from the interwebs, but I was able to get the code running on an Ubuntu Server VM on my Mac.  We need to spend some time nailing down this process for each OS and make sure we have it dialed in.  Getting the updated node.js on Ubuntu was complicated.  Walter and I can handle that task.  I will open it.